requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
pydantic
pysqlite3-binary